refactor: rename AutoStakeMaturity to NnsNnsAutoStakeMaturity
