+++ frontend/src/lib/components/neuron-detail/actions/NnsAutoStakeMaturity.svelte
